+++ source/Energinet.DataHub.MeteringPoints.Application/Providers/MeteringPointOwnership/MeteringPointOwnershipProvider.cs
@@ -1,0 +1,27 @@
+﻿// Copyright 2020 Energinet DataHub A/S
+//
+// Licensed under the Apache License, Version 2.0 (the "License2");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using System.Threading.Tasks;
+
+namespace Energinet.DataHub.MeteringPoints.Application.Providers.MeteringPointOwnership
+{
+    public class MeteringPointOwnershipProvider : IMeteringPointOwnershipProvider
+    {
+        public Task<Owner> GetOwnerAsync(string gsrnNumber)
+        {
+            //TODO: Currently grid area does not hold any details about the owner. Thus, this method cannot be completed. We just return a fake GLN number
+            return Task.FromResult(new Owner("8200000001409"));
+        }
+    }
+}
